+++ Source/RethinkDb.Driver/Net/Connection.cs
@@ -446,12 +446,10 @@ namespace RethinkDb.Driver.Net {
             SendQueryNoReply(Query.Start(NewToken(), term, opts));
         }
 
-        async Task<string> IConnection.RunResultAsRawJson(ReqlAst term, object globalOpts, CancellationToken cancelToken) {
+        Task<Response> IConnection.RunUnsafeAsync(ReqlAst term, object globalOpts, CancellationToken cancelToken)
+        {
             Query q = PrepareQuery(term, OptArgs.FromAnonType(globalOpts));
-            Response res = await SendQuery(q, cancelToken, awaitResponse: true).ConfigureAwait(false);
-            if (res.IsError)
-                throw res.MakeError(q);
-            return res.Data.ToString();
+            return SendQuery(q, cancelToken, awaitResponse: true);
         }
 
         #endregion
@@ -630,7 +628,20 @@ namespace RethinkDb.Driver.Net {
             }
         }
 
+<<<<<<< HEAD
         private static void AssertKeyIsNotBanned(string licenseKey) {
+=======
+        internal static void AssertKeyIsNotBanned(string licenseKey)
+        {
+            var isBanned =
+                licenseKey.StartsWith(
+                    "fuIIq8Pre2NzXVi0otn54PCx22NNAbNReAsk/ylDIV/ZrWeC60B+C76oj3/Ptb8b02vxPYdN6nR2nz3IgYG/O6Zy5TKoYl2UnR2aNq8sKxjv9siwsjMS82EZB8pxs0UwPoz+xmrKY40sqiIz+thDI2EH1MlGoZd+KfJImJp7fvI=")
+                ||
+                licenseKey.StartsWith(
+                    "tE4z+qpOuKWP4XfmAbnyepzI6m/qx2DaI+aDkMes94ujERmA7O6bb0100+LiClLymVLXYXNvkRBg7ot6NGlfyli/8x1h3IgL+HD8gFoWdTAN4oG8wE8ZyrFugnqmAHUDAy4h/KrOqB8VUXwGQh8Y/0ZxOBQb0KOaZJC/MUMbve8=");
+
+            if( isBanned ) throw new UnauthorizedAccessException("The license key has expired.");
+>>>>>>> de65c2550b6852b693c36fb30a19a818c46c9b11
         }
     }
 }